Install official and AUR packages
The package manifests are organized into two files: one for the official
packages located in the Official Arch Linux Repositories (to be
installed by pacman) and one for the PKGBUILDs located in the Arch User
Repositories (to be installed by yay).

Each package name is accompanied by a short description of the package.
Obviously, these files aren't valid inputs to pacman and yay so some
`awk` and `sed` preprocessing magic is performed to transform them into
valid inputs.