Install base packages
vim-tmux-navigator wasn't working. I was so perplexed. I tried disabling
all of the other vim plugins and pieces of configuration that I was
using to see if there was some conflicting configuration. Nope.
vim-tmux-navigator uses `ps` to determine if vim is active -- and -- it
wasn't installed. /facepalm

I can't believe that I forgot to install this package group.