Bring back NERDTree
I was getting an error when opening NERDTree while in an empty vim
buffer. I believe the reason was that NERDTree was trying to
automatically find and select the currently opened file -- except that
the empty buffer doesn't exist in the file tree.

The warning was incredibly annoying, so I removed NERDTree altogether.
But I really like it, so I'm bringing it back. I got the warning to
disappear. All I did was find the line in the stack trace that threw the
error and suppressed it with `silent!`.